+++ common/cnn_common.h
@@ -169,7 +169,8 @@ const uint8_t* get_param_base_pointer(const ParameterInfo *param, uint32_t *limi
 const uint8_t* get_param_row_base_pointer(const ParameterInfo *param, uint32_t *limit_p);
 const uint8_t* get_param_col_base_pointer(const ParameterInfo *param, uint32_t *limit_p);
 const uint8_t* get_param_first_tile_index_base_pointer(const ParameterInfo *param, uint32_t *limit_p);
-int16_t get_q15_param(Model* model, const ParameterInfo *param, uint16_t offset_in_word);
+// FIXME: descript in common/platform.cpp
+int16_t get_q15_param(Model* model, const ParameterInfo *param, uint32_t offset_in_word);
 void put_q15_param(ParameterInfo *param, uint16_t offset_in_word, int16_t val);
 int64_t get_int64_param(const ParameterInfo *param, size_t i);
 uint16_t get_next_slot(Model *model, const ParameterInfo *param);
@@ -177,7 +178,8 @@ const ParameterInfo* get_parameter_info(uint16_t i);
 const Node* get_node(size_t i);
 const Node* get_node(const ParameterInfo* param);
 SlotInfo * get_slot_info(Model* model, uint8_t i);
-void my_memcpy_from_param(Model* model, void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n);
+// FIXME: descript in common/platform.cpp
+void my_memcpy_from_param(Model* model, void *dest, const ParameterInfo *param, uint32_t offset_in_word, size_t n);
 void my_memcpy_from_param_row(Model* model, void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n);
 void my_memcpy_from_param_col(Model* model, void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n);
 void my_memcpy_from_param_first_tile_index(Model* model, void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n);
