+++ common/platform.cpp
@@ -46,7 +46,9 @@ const char* datatype_name<Model>(void) {
     return "model";
 }
 
-void my_memcpy_to_param(ParameterInfo *param, uint16_t offset_in_word, const void *src, size_t n, uint16_t timer_delay) {
+// FIXME: Replce 32 bit to 16 bit for TI compiler
+// The 32 bit is only used to test the correction on server
+void my_memcpy_to_param(ParameterInfo *param, uint32_t offset_in_word, const void *src, size_t n, uint16_t timer_delay) {
     MY_ASSERT(param->bitwidth == 16);
     MY_ASSERT(param->slot < SLOT_CONSTANTS_MIN);
     uint32_t total_offset = param->params_offset + offset_in_word * sizeof(int16_t);
@@ -54,7 +56,8 @@ void my_memcpy_to_param(ParameterInfo *param, uint16_t offset_in_word, const voi
     write_to_nvm(src, intermediate_values_offset(param->slot) + total_offset, n, timer_delay);
 }
 
-void my_memcpy_from_intermediate_values(void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n) {
+// FIXME: Replce 32 bit to 16 bit for TI compiler
+void my_memcpy_from_intermediate_values(void *dest, const ParameterInfo *param, uint32_t offset_in_word, size_t n) {
     read_from_nvm(dest, intermediate_values_offset(param->slot) + offset_in_word * sizeof(int16_t), n);
 }
