+++ common/platform.h
@@ -16,8 +16,10 @@ extern uint8_t dma_counter_enabled;
 
 [[ noreturn ]] void ERROR_OCCURRED(void);
 void my_memcpy(void* dest, const void* src, size_t n);
-void my_memcpy_to_param(ParameterInfo *param, uint16_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
-void my_memcpy_from_intermediate_values(void *dest, const ParameterInfo *param, uint16_t offset_in_word, size_t n);
+// FIXME: descript in .cpp
+void my_memcpy_to_param(ParameterInfo *param, uint32_t offset_in_word, const void *src, size_t n, uint16_t timer_delay);
+// FIXME: descript in .cpp
+void my_memcpy_from_intermediate_values(void *dest, const ParameterInfo *param, uint32_t offset_in_word, size_t n);
 void read_from_samples(void *dest, uint16_t offset_in_word, size_t n);
 ParameterInfo* get_intermediate_parameter_info(uint8_t i);
 void commit_intermediate_parameter_info(uint8_t i);
